Add files via upload
This website is useful for the Indian Railways for getting the information of passenger including their vaccination report, rapid test in the covid-19 period. Django framework is used to build the website. The website consist of registration and login section for the passengers and admin portal where the members of Indian Railway can login and keep track of all the data on the website.